feat(model): bf16 default precision end to end (ROADMAP F5)
Checkpoint torch_dtype now drives ModelConfig.dtype (bf16 when undeclared,
fp32 downgraded); the loader materialises parameters and widens FP8 weights
in that dtype instead of blanket fp16. Triton kernels (flash attention,
flash decoding, w4a16/w8a16/w8a8/fp8 linear) accept bf16 activations --
tl.dot operands follow the activation dtype, accumulation stays fp32.
fp16 checkpoints keep fp16 parameters. Golden baselines regenerated in
bf16 (Qwen2.5-0.5B, Qwen3-0.6B, Qwen3-0.6B-FP8).

+++ tests/models/test_qwen3_moe.py
@@ -55,7 +55,8 @@ def test_block_fp8_dequant(shape):
 
     deq = dequant_block_fp8(w8, scale_inv)
 
-    assert deq.dtype == torch.float16
+    # bf16 is the loader's default widening target since v0.9 (ROADMAP F5).
+    assert deq.dtype == torch.bfloat16
     assert deq.shape == w.shape
     # e4m3 keeps ~3 mantissa bits -> ~6% relative error per element.
     rel = ((deq.float() - w).abs() / w.abs().clamp_min(1e-3)).median()
